Write MPV version to registry and upate if necessary
It would be preferable to read the EXE version directly but it doesn't use a sequential version so we assume whatever the installer wrote last is what is currently installed. The absence of the version string in the registry will cause existing installations to reinstall MPV and YTDL even if they are up to date.


git-svn-id: https://subversion.assembla.com/svn/smplayer/smplayer/trunk@7168 fd4b1137-b938-0410-b384-8fdad0efb714